# docs/api/openapi.yaml
components:
  schemas:
...
    DmarcImportResponse:
      additionalProperties: false
      description: Idempotent DMARC import result.
      properties:
        record_count:
          minimum: 0.0
          title: Record Count
          type: integer
        status:
          enum:
          - imported
          - already_imported
          - ignored
          title: Status
          type: string
      required:
      - status
      - record_count
      title: DmarcImportResponse
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ReportAddressResponse:
      additionalProperties: false
      description: Signed false-negative forwarding address.
      properties:
        address:
          title: Address
          type: string
      required:
      - address
      title: ReportAddressResponse
      type: object
    ReportedEmailIngestResponse:
      additionalProperties: false
      description: Idempotent forwarded-message ingestion result.
      properties:
        idempotent:
          title: Idempotent
          type: boolean
        status:
          const: accepted
          title: Status
          type: string
      required:
      - status
      - idempotent
      title: ReportedEmailIngestResponse
      type: object
    ReportedEmailListResponse:
      additionalProperties: false
      description: Forwarded reports belonging to the authenticated workspace.
      properties:
        items:
          items:
            $ref: '#/components/schemas/ReportedEmailSummary'
          title: Items
          type: array
      required:
      - items
      title: ReportedEmailListResponse
      type: object
    ReportedEmailSummary:
      additionalProperties: false
      description: 'One forwarded report, described without reproducing its content.


        Deliberately metadata only. The ingest pipeline anonymises the message and

        stores it in private R2 precisely so the forwarded content stops circulating;

        returning a body here would undo that. Date, size and status are enough for

        the person who forwarded it to see that it arrived.'
      properties:
        id:
          title: Id
          type: string
        received_at:
          title: Received At
          type: string
        size_bytes:
          title: Size Bytes
          type: integer
        status:
          title: Status
          type: string
      required:
      - id
      - received_at
      - size_bytes
      - status
      title: ReportedEmailSummary
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  description: Generated from the deployed FastAPI routes and Pydantic models. Authenticated customer
    operations use Better Auth sessions or bearer tokens; internal and email-gateway routes require their
    documented service credentials.
  summary: Data, application, and integration contracts for Sicurre.
  title: Sicurre Platform API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /v1/email/dmarc-reports:
    post:
      description: Accept one Worker-authenticated aggregate DMARC report email.
      operationId: ingest_dmarc_email_v1_email_dmarc_reports_post
      parameters:
      - in: header
        name: x-sicurre-report-key
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Sicurre-Report-Key
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DmarcImportResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Ingest Dmarc Email
      tags:
      - reported-email
...
  /v1/email/reports/{token}:
    post:
      description: Accept one Worker-authenticated forwarded message as sanitized evidence.
      operationId: ingest_reported_email_v1_email_reports__token__post
      parameters:
      - in: path
        name: token
        required: true
        schema:
          title: Token
          type: string
      - in: header
        name: x-sicurre-report-key
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Sicurre-Report-Key
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportedEmailIngestResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Ingest Reported Email
      tags:
      - reported-email
...
  /v1/feedback/report-address:
    get:
      description: Return the signed forwarding alias for the authenticated workspace.
      operationId: get_report_address_v1_feedback_report_address_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportAddressResponse'
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: Get Report Address
      tags:
      - reported-email
  /v1/feedback/reports:
    get:
      description: 'List forwarded reports for the authenticated workspace.


        Returns metadata only; the message content stays in private R2. Scoped

        on ``workspace_id`` like every other tenant route.'
      operationId: list_reported_emails_v1_feedback_reports_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportedEmailListResponse'
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: List Reported Emails
      tags:
      - reported-email
...
servers:
- description: Current Sicurre deployment
  url: /
...